backend/requirements.txt
fastapi
uvicorn
pandas
scipy
google-generativeai
firebase-admin
python-multipart
pydantic
python-dotenv
jinja2
